Added improved Dijkstra, animation in GraphicController and better error communication in InputFileReader

+++ src/main/java/Hospital.java
@@ -71,7 +71,7 @@ public class Hospital extends MapObject {
     }
 
 
-    public void setDistance(int distance) {
+    public void setDistance(double distance) {
         this.distance = distance;
     }
